[FileDataManager] Rewrite interface hierarhy and file management
